r-car: boot with ro, not rw
Booting with ro is safer in the case of filesystem corruption because it
prevents processes from writing to a broken filesystem before the kernel
has a chance to fsck and then remount rw. For this reason, it seems to
be the standard in other Linux distros. From my testing, I don't see any
drawback to it.

Signed-off-by: Martin Kelly <mkelly@xevo.com>