Containers: Rename of the MD file for Containers plus edits
I renamed the old SDK Quick Start file to be a
"docker-container-setup.md" file.  This is more appropriate
for the content of this file.

Edits included a complete re-write of the content to result
in a procedure for installing Docker and setting up a container
that has an installed SDK.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>