Fixed links in "Debug the Application" section.
The links in this section need to resolve to new files in the
XDS section.  I updated the links.  There could be some overlap
where the links do not work if a review for the xds-docs repo
is not merged.  That outstanding review is for the renaming of
the files.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>