app-workflow-build-app: Add info about linux-oe-* for qmake
When building a Qt qmake project on desktop or AGL, this can be used to do
something different when compiling outside of the AGL environment vs. when
building for AGL through bitbake.

Unsure what happens when building the application through the SDK, may be
necessary to adjust this tip to work under more scenarios.

I found this QMAKE_SPEC by running a devshell for the app recipe, then running
`qmake -query`